requirements.txt
lancedb>=0.3.1
cognee>=0.1.15
openai>=1.3.0
streamlit>=1.28.0
pypdf2>=3.0.1
python-docx>=0.8.11
pandas>=2.0.0
numpy>=1.24.0
sentence-transformers>=2.2.2
langchain>=0.1.0
langchain-openai>=0.0.5
python-dotenv>=1.0.0
tiktoken>=0.5.0
faiss-cpu>=1.7.4 